application/requirements.txt
amqp==5.0.5
appdirs==1.4.4
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
ffmpeg-python==0.2.0
future==0.18.2
kombu==5.0.2
mypy-extensions==0.4.3
pathspec==0.8.1
prompt-toolkit==3.0.18
pyee==8.1.0
pytz==2021.1
regex==2021.3.17
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
vine==5.0.0
wcwidth==0.2.5
mysqlclient